Predicting the Best Fit Employees to a Project
Talent management and strategic HR are the vital factors of Human Resource
Management (HRM). The technology evolution leads to the demand for skilled and
knowledgeable individuals with the ever-changing requirement for product and services.
HRM needs to manage the expertise and talent in challenging enterprises. It is
challenging to allocate the skilled employee for a project by manually screening their
resume due to time constraints. Maintaining employee satisfaction is the perennial
challenge for the company. The approach outlined in this research will help to utilise
employee’s details along with keeping budgetary constraints into mind, allocating best
employees for the project and finding the attrition pattern to retain the crucial employees.
Additionally, finding the low performers and re-skill them to their target skills is also
necessary. The Predicting Best Employee Selection (PBES) solution will provide the
resolution of all the complications in that by entering the budgetary constraints, HRM
team or Manager can get a list of best-fit employees to choose. Then, the list of the low
performers can help the HRM team to assign training on the domain. Furthermore, it
helps in identifying the attrition pattern to retain the crucial players. The employee’s
grading clusters achieved successfully with K-mean model along with completion of
SVM model will successful find the attrition pattern. The Integration with existing HR
application remains uncovered.